requirements.txt
opencv-python
mediapipe==0.10.21
numpy
scipy
ultralytics
pyrealsense2
PyQt5
torch
transformers>=5.8.0
tokenizers>=0.22.2
huggingface_hub>=1.13.0
accelerate
bitsandbytes
Pillow
jinja2>=3.1.0
num2words

# Optional for real GraspNet inference:
# pip install -r requirements-graspnet.txt
